React/Image-resize-application/backend/requirements.txt
fastapi
uvicorn
pillow
tweepy
python-multipart